+++ clients/odbc/driver/SQLTablePrivileges.c
@@ -58,7 +58,10 @@ SQLTablePrivileges_(ODBCStmt *stmt,
 	fixODBCstring(szTableName, nTableNameLength, SQLSMALLINT, addStmtError, stmt, return SQL_ERROR);
 
 #ifdef ODBCDEBUG
-	ODBCLOG("\"%.*s\" \"%.*s\" \"%.*s\"\n", nCatalogNameLength, (char*)szCatalogName, nSchemaNameLength, (char*)szSchemaName, nTableNameLength, (char*)szTableName);
+	ODBCLOG("\"%.*s\" \"%.*s\" \"%.*s\"\n",
+		(int) nCatalogNameLength, (char *) szCatalogName,
+		(int) nSchemaNameLength, (char *) szSchemaName,
+		(int) nTableNameLength, (char *) szTableName);
 #endif
 
 	/* SQLTablePrivileges returns a table with the following columns:
